Extract TestProject
